Split controllers into one file per function
