+++ apps/api/src/controllers/auth/index.ts
@@ -1,0 +1,15 @@
+import login from './login';
+import logout from './logout';
+import signup from './signup';
+import verify from './verify';
+import resendVerification from './resend-verification';
+
+const auth = {
+  login,
+  logout,
+  signup,
+  verify,
+  resendVerification,
+};
+
+export default auth;
