+++ apps/api/src/controllers/index.ts
@@ -1,5 +1,5 @@
-import * as auth from './auth';
-import * as user from './user';
+import auth from './auth';
+import user from './user';
 
 const controllers = {
   auth,
