+++ apps/api/src/controllers/user/index.ts
@@ -1,0 +1,7 @@
+import me from './me';
+
+const user = {
+  me,
+};
+
+export default user;
