+++ package.json
@@ -38,6 +38,7 @@
     "cors": "^2.8.5",
     "dayjs": "^1.11.13",
     "express": "^4.21.2",
+    "http-status-codes": "^2.3.0",
     "jsonwebtoken": "^9.0.2",
     "next": "~15.1.4",
     "nodemailer": "^6.10.0",
